actuator: Fix active low config
Before this commit it was not possible to configure active low
actuator, see #30.

This commit changes the boolean Kconfig symbol
ACTUATOR_ACTIVE_HIGH (where CONFIG_ACTUATOR_ACTIVE_HIGH is not defined
for active low config) to the integer Kconfig value
ACTUATOR_ACTIVE_LEVEL, which is always defined, either to 1 or to 0.
This way there is no need for ifdefs in the c++ code.

This fixes #30 and makes #32 obsolete.